@@ -1,6 +1,6 @@
 module github.com/status-im/status-keycard-go
 
-go 1.16
+go 1.22
 
 require (
 	github.com/ebfe/scard v0.0.0-20241214075232-7af069cabc25
@@ -8,4 +8,11 @@ require (
 	github.com/status-im/keycard-go v0.3.3
 	golang.org/x/crypto v0.1.0
 	golang.org/x/text v0.4.0
+)
+
+require (
+	github.com/btcsuite/btcd/btcec/v2 v2.2.0 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1 // indirect
+	github.com/go-stack/stack v1.8.1 // indirect
+	golang.org/x/sys v0.2.0 // indirect
 )
