@@ -9,8 +9,8 @@ import (
 	"errors"
 	"unsafe"
 
-	"github.com/status-im/status-keycard-go/signal"
 	"github.com/status-im/status-keycard-go/pkg/flow"
+	"github.com/status-im/status-keycard-go/signal"
 )
 
 func main() {}
